Query CRAN(like) package data

It uses CRAN and BioConductor packages, for the current platform and R version, from the default repositories.

Usage

meta_cache_deps(packages, dependencies = NA, recursive = TRUE)

meta_cache_revdeps(packages, dependencies = NA, recursive = TRUE)

meta_cache_update()

meta_cache_list(packages = NULL)

meta_cache_cleanup(force = FALSE)

meta_cache_summary()

Arguments

packages

Packages to query.

dependencies

Dependency types to query. See the dependencies parameter of utils::install.packages().

recursive

Whether to query recursive dependencies.

force

Whether to force cleanup without asking the user.

Value

A data frame of the dependencies. For meta_cache_deps() and meta_cache_revdeps() it includes the queried packages as well.

Details

meta_cache_list() lists all packages.

meta_cache_update() updates all metadata. Note that metadata is automatically updated if it is older than seven days.

meta_cache_deps() queries packages dependencies.

meta_cache_revdeps() queries reverse package dependencies.

meta_cache_summary() lists data about the cache, including its location and size.

meta_cache_cleanup() deletes the cache files from the disk.

Examples

meta_cache_list("pkgdown")
#> 
#> ✔ Updated metadata database: 969.23 kB in 5 files.
#> 
#> ℹ Updating metadata database
#> ✔ Updating metadata database ... done
#> 
#> # A data frame: 2 × 32
#>   package version depends  sugge…¹ license needs…² imports linki…³ enhan…⁴
#> * <chr>   <chr>   <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 pkgdown 2.0.6   R (>= 3… "covr,… MIT + … no      "bslib… NA      NA     
#> 2 pkgdown 2.0.6   R (>= 3… "covr,… MIT + … no      "bslib… NA      NA     
#> # … with 23 more variables: os_type <chr>, path <chr>, repodir <chr>,
#> #   rversion <chr>, platform <chr>, priority <chr>, ref <chr>,
#> #   type <chr>, direct <lgl>, status <chr>, target <chr>, mirror <chr>,
#> #   sources <list>, filesize <int>, sha256 <chr>, sysreqs <chr>,
#> #   built <chr>, published <dttm>, deps <list>, md5sum <chr>,
#> #   license_restricts_use <chr>, license_is_foss <chr>, archs <chr>, and
#> #   abbreviated variable names ¹​suggests, ²​needscompilation, …
meta_cache_deps("pkgdown", recursive = FALSE)
#> # A data frame: 42 × 32
#>    package version depends sugge…¹ license needs…² imports linki…³ enhan…⁴
#>  * <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 bslib   0.4.1   R (>= … shiny … MIT + … no      grDevi… NA      NA     
#>  2 callr   3.7.3   R (>= … asciic… MIT + … no      proces… NA      NA     
#>  3 cli     3.4.1   R (>= … callr,… MIT + … yes     utils   NA      NA     
#>  4 desc    1.4.2   R (>= … callr,… MIT + … no      cli, R… NA      NA     
#>  5 digest  0.6.30  R (>= … tinyte… GPL (>… yes     utils   NA      NA     
#>  6 downlit 0.4.2   R (>= … covr, … MIT + … no      brio, … NA      NA     
#>  7 fs      1.5.2   R (>= … testth… MIT + … yes     methods NA      NA     
#>  8 httr    1.4.4   R (>= … covr, … MIT + … no      curl (… NA      NA     
#>  9 jsonli… 1.8.3   methods httr, … MIT + … yes     NA      NA      NA     
#> 10 magrit… 2.0.3   R (>= … covr, … MIT + … yes     NA      NA      NA     
#> # … with 32 more rows, 23 more variables: os_type <chr>, path <chr>,
#> #   repodir <chr>, rversion <chr>, platform <chr>, priority <chr>,
#> #   ref <chr>, type <chr>, direct <lgl>, status <chr>, target <chr>,
#> #   mirror <chr>, sources <list>, filesize <int>, sha256 <chr>,
#> #   sysreqs <chr>, built <chr>, published <dttm>, deps <list>,
#> #   md5sum <chr>, license_restricts_use <chr>, license_is_foss <chr>,
#> #   archs <chr>, and abbreviated variable names ¹​suggests, …
meta_cache_revdeps("pkgdown", recursive = FALSE)
#> # A data frame: 8 × 32
#>   package  version depends sugge…¹ license needs…² imports linki…³ enhan…⁴
#> * <chr>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 CALANGO  1.0.12  R (>= … "Annot… GPL-2   no      "asser… NA      NA     
#> 2 devtools 2.4.5   R (>= … "BiocM… MIT + … no      "cli (… NA      NA     
#> 3 pkgdown  2.0.6   R (>= … "covr,… MIT + … no      "bslib… NA      NA     
#> 4 prefera… 0.4.1   R (>= …  NA     MIT + … no       NA     NA      NA     
#> 5 CALANGO  1.0.12  R (>= … "Annot… GPL-2   no      "asser… NA      NA     
#> 6 devtools 2.4.5   R (>= … "BiocM… MIT + … no      "cli (… NA      NA     
#> 7 pkgdown  2.0.6   R (>= … "covr,… MIT + … no      "bslib… NA      NA     
#> 8 prefera… 0.4.1   R (>= …  NA     MIT + … no       NA     NA      NA     
#> # … with 23 more variables: os_type <chr>, path <chr>, repodir <chr>,
#> #   rversion <chr>, platform <chr>, priority <chr>, ref <chr>,
#> #   type <chr>, direct <lgl>, status <chr>, target <chr>, mirror <chr>,
#> #   sources <list>, filesize <int>, sha256 <chr>, sysreqs <chr>,
#> #   built <chr>, published <dttm>, deps <list>, md5sum <chr>,
#> #   license_restricts_use <chr>, license_is_foss <chr>, archs <chr>, and
#> #   abbreviated variable names ¹​suggests, ²​needscompilation, …