Query CRAN(like) package data

It uses CRAN and BioConductor packages, for the current platform and R version, from the default repositories.

Usage

meta_cache_deps(packages, dependencies = NA, recursive = TRUE)

meta_cache_revdeps(packages, dependencies = NA, recursive = TRUE)

meta_cache_update()

meta_cache_list(packages = NULL)

meta_cache_cleanup(force = FALSE)

meta_cache_summary()

Arguments

packages

Packages to query.

dependencies

Dependency types to query. See the dependencies parameter of utils::install.packages().

recursive

Whether to query recursive dependencies.

force

Whether to force cleanup without asking the user.

Value

A data frame of the dependencies. For meta_cache_deps() and meta_cache_revdeps() it includes the queried packages as well.

Details

meta_cache_list() lists all packages.

meta_cache_update() updates all metadata. Note that metadata is automatically updated if it is older than seven days.

meta_cache_deps() queries packages dependencies.

meta_cache_revdeps() queries reverse package dependencies.

meta_cache_summary() lists data about the cache, including its location and size.

meta_cache_cleanup() deletes the cache files from the disk.

Examples

meta_cache_list("pkgdown")
#> ℹ Loading metadata database
#> ✔ Loading metadata database ... done
#> 
#> # A data frame: 2 × 32
#>   package version suggests     needscompilation license depends imports
#> * <chr>   <chr>   <chr>        <chr>            <chr>   <chr>   <chr>  
#> 1 pkgdown 2.1.3   "covr, diff… NA               MIT + … R (>= … "bslib…
#> 2 pkgdown 2.1.3   "covr, diff… no               MIT + … R (>= … "bslib…
#> # ℹ 25 more variables: linkingto <chr>, enhances <chr>,
#> #   license_restricts_use <chr>, os_type <chr>, path <chr>,
#> #   priority <chr>, license_is_foss <chr>, archs <chr>, repodir <chr>,
#> #   rversion <chr>, platform <chr>, ref <chr>, type <chr>,
#> #   direct <lgl>, status <chr>, target <chr>, mirror <chr>,
#> #   sources <list>, filesize <dbl>, sha256 <chr>, sysreqs <chr>,
#> #   built <chr>, published <dttm>, deps <list>, md5sum <chr>
meta_cache_deps("pkgdown", recursive = FALSE)
#> # A data frame: 40 × 32
#>    package    version suggests needscompilation license depends imports
#>  * <chr>      <chr>   <chr>    <chr>            <chr>   <chr>   <chr>  
#>  1 bslib      0.9.0   "bsicon… NA               MIT + … R (>= … "base6…
#>  2 callr      3.7.6   "asciic… NA               MIT + … R (>= … "proce…
#>  3 cli        3.6.5   "callr,… NA               MIT + … R (>= … "utils"
#>  4 desc       1.4.3   "callr,… NA               MIT + … R (>= … "cli, …
#>  5 downlit    0.4.4   "covr, … NA               MIT + … R (>= … "brio,…
#>  6 fontaweso… 0.5.3   "covr, … NA               MIT + … R (>= … "rlang…
#>  7 fs         1.6.6   "covr, … NA               MIT + … R (>= … "metho…
#>  8 httr2      1.2.1   "askpas… NA               MIT + … R (>= … "cli (…
#>  9 jsonlite   2.0.0   "httr, … NA               MIT + … methods  NA    
#> 10 openssl    2.3.4   "curl, … NA               MIT + … NA      "askpa…
#> # ℹ 30 more rows
#> # ℹ 25 more variables: linkingto <chr>, enhances <chr>,
#> #   license_restricts_use <chr>, os_type <chr>, path <chr>,
#> #   priority <chr>, license_is_foss <chr>, archs <chr>, repodir <chr>,
#> #   rversion <chr>, platform <chr>, ref <chr>, type <chr>,
#> #   direct <lgl>, status <chr>, target <chr>, mirror <chr>,
#> #   sources <list>, filesize <dbl>, sha256 <chr>, sysreqs <chr>, …
meta_cache_revdeps("pkgdown", recursive = FALSE)
#> # A data frame: 10 × 32
#>    package    version suggests needscompilation license depends imports
#>  * <chr>      <chr>   <chr>    <chr>            <chr>   <chr>   <chr>  
#>  1 CALANGO    1.0.20  "Annota… NA               GPL-2   R (>= … "asser…
#>  2 devtools   2.4.6   "BiocMa… NA               MIT + … R (>= … "cli (…
#>  3 pkgdown    2.1.3   "covr, … NA               MIT + … R (>= … "bslib…
#>  4 pkgdown.o… 0.1.1   "knitr,… NA               MIT + … NA      "pkgdo…
#>  5 washr      1.0.1   "knitr,… NA               GPL (>… NA      "desc …
#>  6 CALANGO    1.0.20  "Annota… no               GPL-2   R (>= … "asser…
#>  7 devtools   2.4.6   "BiocMa… no               MIT + … R (>= … "cli (…
#>  8 pkgdown    2.1.3   "covr, … no               MIT + … R (>= … "bslib…
#>  9 pkgdown.o… 0.1.1   "knitr,… no               MIT + … NA      "pkgdo…
#> 10 washr      1.0.1   "knitr,… no               GPL (>… NA      "desc …
#> # ℹ 25 more variables: linkingto <chr>, enhances <chr>,
#> #   license_restricts_use <chr>, os_type <chr>, path <chr>,
#> #   priority <chr>, license_is_foss <chr>, archs <chr>, repodir <chr>,
#> #   rversion <chr>, platform <chr>, ref <chr>, type <chr>,
#> #   direct <lgl>, status <chr>, target <chr>, mirror <chr>,
#> #   sources <list>, filesize <dbl>, sha256 <chr>, sysreqs <chr>,
#> #   built <chr>, published <dttm>, deps <list>, md5sum <chr>