Show the status of CRAN-like repositories

It checks the status of the configured or supplied repositories, for the specified platforms and R versions.

Usage

repo_status(
  platforms = default_platforms(),
  r_version = getRversion(),
  bioc = TRUE,
  cran_mirror = default_cran_mirror()
)

Arguments

platforms

Platforms to use, default is default_platforms().

r_version

R version(s) to use, the default is the current R version, via getRversion().

bioc

Whether to add the Bioconductor repositories. If you already configured them via options(repos), then you can set this to FALSE. See bioc_version() for the details about how pkgcache handles Bioconductor repositories.

cran_mirror

The CRAN mirror to use, see default_cran_mirror().

Value

A data frame that has a row for every repository, on every queried platform and R version. It has these columns:

name: the name of the repository. This comes from the names of the configured repositories in options("repos"), or added by pkgcache. It is typically CRAN for CRAN, and the current Bioconductor repositories are BioCsoft, BioCann, BioCexp, BioCworkflows, BioCbooks.

url: base URL of the repository.

bioc_version: Bioconductor version, or NA for non-Bioconductor repositories.

platform: platform, see default_platforms() for possible values.

path: the path to the packages within the base URL, for a given platform and R version.

r_version: R version, one of the specified R versions.

ok: Logical flag, whether the repository contains a metadata file for the given platform and R version.

ping: HTTP response time of the repository in seconds. If the ok column is FALSE, then this columns in NA.

error: the error object if the HTTP query failed for this repository, platform and R version.

Details

The returned data frame has a summary() method, which shows the same information is a concise table. See examples below.

See also

Examples

repo_status()
#> # A data frame: 7 × 10
#>   name       url   type  bioc_version platform path  r_version ok     ping
#>   <chr>      <chr> <chr> <chr>        <chr>    <chr> <chr>     <lgl> <dbl>
#> 1 RSPM       http… cran… NA           source   src/… 4.3       TRUE  0.836
#> 2 CRAN       http… cran  NA           source   src/… 4.3       TRUE  0.831
#> 3 BioCsoft   http… bioc  3.18         source   src/… 4.3       TRUE  0.830
#> 4 BioCann    http… bioc  3.18         source   src/… 4.3       TRUE  0.830
#> 5 BioCexp    http… bioc  3.18         source   src/… 4.3       TRUE  0.833
#> 6 BioCworkf… http… bioc  3.18         source   src/… 4.3       TRUE  0.833
#> 7 BioCbooks  http… bioc  3.18         source   src/… 4.3       TRUE  0.836
#> # ℹ 1 more variable: error <list>
rst <- repo_status(
  platforms = c("windows", "macos"),
  r_version = c("4.0", "4.1")
)
summary(rst)
#> Repository summary:                               i386+x86_64-w64-mingw32 x86_64-apple-darwin17.0         
#> RSPM          @ packagemanager.posit.co (R 4.0)              ✔                       ✔              (1.4s)
#> CRAN          @ cran.rstudio.com        (R 4.0)              ✔                       ✔              (1.2s)
#> RSPM          @ packagemanager.posit.co (R 4.1)              ✔                       ✔              (1.4s)
#> CRAN          @ cran.rstudio.com        (R 4.1)              ✔                       ✔              (1.2s)
#> BioCsoft      @ bioconductor.org        (R 4.0)              ✔                       ✔              (1.9s)
#> BioCann       @ bioconductor.org        (R 4.0)              ✔                       ✔              (2.1s)
#> BioCexp       @ bioconductor.org        (R 4.0)              ✔                       ✔              (2s  )
#> BioCworkflows @ bioconductor.org        (R 4.0)              ✔                       ✔              (2.2s)
#> BioCbooks     @ bioconductor.org        (R 4.0)              ✔                       ✔              (1.3s)
#> BioCsoft      @ bioconductor.org        (R 4.1)              ✔                       ✔              (2s  )
#> BioCann       @ bioconductor.org        (R 4.1)              ✔                       ✔              (2s  )
#> BioCexp       @ bioconductor.org        (R 4.1)              ✔                       ✔              (2s  )
#> BioCworkflows @ bioconductor.org        (R 4.1)              ✔                       ✔              (2s  )
#> BioCbooks     @ bioconductor.org        (R 4.1)              ✔                       ✔              (1.3s)